Edit RecipeContainer and move it to Components folder. Delete Container folder.
